requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.2.0
confection==0.1.4
contourpy==1.2.0
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.7
django-countries==7.5.1
django-import-export==3.3.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.45.1
freetype-py==2.3.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
html5lib==1.1
idna==3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
keras-models==0.0.7
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
lxml==4.9.4
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
MouseInfo==0.1.3
mss==9.0.1
murmurhash==1.0.10
mysql-connector-python==8.2.0
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.0.0
preshed==3.0.9
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyGetWindow==0.0.9
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyMsgBox==1.0.9
pyparsing==3.1.1
pypdf==3.17.3
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
python-bidi==0.4.2
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.10.3
reportlab==4.0.8
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.3.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
sqlparse==0.4.4
srsly==2.4.8
svglib==1.5.1
tablib==3.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.24.7
wasabi==1.1.2
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.14.1
xhtml2pdf==0.2.13
xlrd==2.0.1
xlwt==1.3.0
